requirements.txt
# 環境變數管理（如果你有使用 .env）
python-dotenv>=1.1.1

# 音訊處理
pydub>=0.25.1

# Whisper 模型（git 最新版）
openai-whisper @ git+https://github.com/openai/whisper.git

# 數值運算套件
numpy==1.26.4

# 運算核心（CPU/GPU）
torch>=2.2.2

# 終端機文字介面的進度條工具
tqdm>=4.67.1


# Gemini
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15
google-api-core==2.26.0
google-auth==2.41.1
